+++ tests/testthat/test_superbToWide.R
@@ -2,7 +2,9 @@ context("Testing suberbToWide")
 
 
 test_that("TESTS (1/4)", {
-    options(superb.feedback = 'none')
+	old <- options() 
+	on.exit(options(old)) 
+    options("superb.feedback" = 'none')
 
 	names(Orange) <- c("Tree","age","DV")
 	Owide <- superbToWide(Orange, id = "Tree", WSFactors = "age", variable = "DV" )
@@ -12,12 +14,14 @@ test_that("TESTS (1/4)", {
     expect_equal( Owide$DV_1582[5], 177)
 
     # restores default information
-    options(superb.feedback = c('design','warnings','summary'))
+    options("superb.feedback" = c('design','warnings','summary'))
 })
 
 
 test_that("TESTS (2/4)", {
-    options(superb.feedback = 'none')
+	old <- options() 
+	on.exit(options(old)) 
+    options("superb.feedback" = 'none')
 
 	ss <- 3
 	dta <- GRD( BSFactors = "Moment (2)", SubjectsPerGroup = ss)
@@ -29,12 +33,14 @@ test_that("TESTS (2/4)", {
     expect_equal( dta$DV[ss*2], tt$DV_2[ss])
 
     # restores default information
-    options(superb.feedback = c('design','warnings','summary'))
+    options("superb.feedback" = c('design','warnings','summary'))
 })
 
 
 test_that("TESTS (3/4)", {
-    options(superb.feedback = 'none')
+	old <- options() 
+	on.exit(options(old)) 
+    options("superb.feedback" = 'none')
 
 	ss <- 7
 	dta <- GRD( BSFactors = c("Moment (2)","time(3)","what(5)"), SubjectsPerGroup = ss)
@@ -46,11 +52,13 @@ test_that("TESTS (3/4)", {
     expect_equal( dta$DV[ss*2*3*5], tt$DV_2_3_5[ss])
 
     # restores default information
-    options(superb.feedback = c('design','warnings','summary'))
+    options("superb.feedback" = c('design','warnings','summary'))
 })
 
 test_that("TESTS (3/4)", {
-    options(superb.feedback = 'none')
+	old <- options() 
+	on.exit(options(old)) 
+    options("superb.feedback" = 'none')
 
 	ss <-13
 	dta <- GRD( BSFactors = c("Moment (2)","time(3)","what(5)","bis(7)"), SubjectsPerGroup = ss)
@@ -62,5 +70,5 @@ test_that("TESTS (3/4)", {
     expect_equal( dta$DV[ss*2*3*5*7], tt$DV_2_3_5_7[ss])
 
     # restores default information
-    options(superb.feedback = c('design','warnings','summary'))
+    options("superb.feedback" = c('design','warnings','summary'))
 })