+++ NEWS.md
@@ -1,3 +1,19 @@
+# superb 0.95.202402 (February 8th, 2024)
+
+* Changed numbering of sub-sub version according to year-month
+* Added a new plot layout, boxplot, to display box plots.
+* Adjusted testthat tests following new release of ggplot2
+* Added unitaryAlpha technique to compute average correlation
+* Simplified documentation for within-subject designs.
+
+# superb 0.95.83 (November 11th, 2023)
+
+* Made a message shown based on option
+
+# superb 0.95.82 (August 19th, 2023)
+
+* Corrected labeling of factors.
+
 # superb 0.95.81 (August 13th, 2023)
 
 * Updated CustomizingSuperbPlots vignettes.